// src/app/generated/model/lifeCycleState.ts
/**
 * document-management
 * 1000kit document management
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export enum LifeCycleState {
  ARCHIVED = 'ARCHIVED',
  DRAFT = 'DRAFT',
  RELEASED = 'RELEASED',
  REVIEW = 'REVIEW',
}